api/requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scipy>=1.13,<2.0
python-multipart>=0.0.9
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib>=1.7,<2.0
alembic>=1.13,<2.0
slowapi>=0.1.9,<1.0
httpx>=0.27,<1.0
structlog>=24.1,<26.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
boto3>=1.34,<2.0
